requirements.txt
pyTelegramBotAPI>=4.12.0
tinydb>=4.7.1
python-dotenv>=1.0.0
pillow>=9.5.0 